# HPC-10029: Optimize `toCamelCase`