In the ticket, it was reported that call to HPC API endpoint `v2/plan/1185/responseMonitoring?includeCaseloadDisaggregation=true&includeIndicatorDisaggregation=false&disaggregationOnlyTotal=false` doesn't complete until timeout is reached. After investigation using profiling tool of DevTools, I found that calls to `toCamelCase` are actually responsible to great extent.

First, I have tried various ways of optimizing the method itself, to only have managed getting from ~45 seconds to ~25 seconds while processing attachment data on HPC API side. Without calls to `toCamelCase`, that code took 5 seconds. Even Lodash's implementation was worse than this, at around 35 seconds, all measured on local. So, just optimization didn't cut it, so I decided to "cache" the responses of this method, which brought down number of its invocations from 16,850,784 to just 4,584! This huge reduction helped tremendously with performance.